WIP pull NimbusRecipe into its own class
