+++ __tests__/lib/nimbusRecipes.test.ts
@@ -1,0 +1,7 @@
+import { NimbusRecipe } from '../../lib/nimbusRecipe'
+
+describe('NimbusRecipe', () => {
+  it('creates a NimbusRecipe from a raw JS recipe object', () => {
+
+  })
+})
